requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.1
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
python-decouple==3.8
requests==2.28.1
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0